refactor: move getExistingPr function into send-to-sns
