+++ packages/dependency-graph-integrator/src/pull-requests.ts
@@ -1,5 +1,4 @@
 import { randomBytes } from 'crypto';
-import type { Endpoints } from '@octokit/types';
 import { stageAwareOctokit } from 'common/src/functions';
 import { addPrToProject } from 'common/src/projects-graphql';
 import type { Octokit } from 'octokit';
@@ -91,36 +90,8 @@ export async function requestTeamReview(
 	}
 }
 
-type PullRequestParameters =
-	Endpoints['GET /repos/{owner}/{repo}/pulls']['parameters'];
 
-type PullRequest =
-	Endpoints['GET /repos/{owner}/{repo}/pulls']['response']['data'][number];
 
-function isGithubAuthor(pull: PullRequest, author: string) {
-	return pull.user?.login === author && pull.user.type === 'Bot';
-}
-
-export async function getExistingPullRequest(
-	octokit: Octokit,
-	repoName: string,
-	owner: string,
-	author: string,
-) {
-	const pulls = await octokit.paginate(octokit.rest.pulls.list, {
-		owner,
-		repo: repoName,
-		state: 'open',
-	} satisfies PullRequestParameters);
-
-	const found = pulls.filter((pull) => isGithubAuthor(pull, author));
-
-	if (found.length > 1) {
-		console.warn(`More than one PR found on ${repoName} - choosing the first.`);
-	}
-
-	return found[0];
-}
 
 export async function createPrAndAddToProject(
 	stage: string,
